Get rid of the slide and synth model for a move
A movement can't be decomposed in a slide and a synthesis
anymore. They are intertwined because a traslation may depend on a
syntehsis being made.

This model gave us a way to differentiate between an attempt to make a
move (determined by the return value of the move function) and
actually making a move on the board (by calling the slide-synth
function). Because we still need this distinction, some function names
have been changed.